requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
python-dotenv==0.19.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0
python-multipart==0.0.5
pydantic==1.8.2
requests==2.26.0
aiohttp==3.8.1
python-jose==3.3.0
passlib==1.7.4 